requirements.txt
# ============================================================
# Bear Detector — Python Dependencies
# Python >= 3.10
# ============================================================

# ── Core ML & Computer Vision ───────────────────────────────
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.1.0          # YOLOv8 detection + segmentation
opencv-python>=4.9.0        # video / image processing
Pillow>=10.0.0              # image I/O

# ── Deep Learning Utilities ─────────────────────────────────
numpy>=1.24.0
scipy>=1.11.0               # Hungarian algorithm in SORT tracker
scikit-learn>=1.3.0         # evaluation metrics (PR curve, confusion matrix)

# ── Tracking ────────────────────────────────────────────────
filterpy>=1.4.5             # Kalman filter for SORT tracker

# ── Configuration & Experiment Tracking ─────────────────────
pyyaml>=6.0.1               # YAML config loading
mlflow>=2.10.0              # experiment tracking

# ── Visualisation ───────────────────────────────────────────
matplotlib>=3.7.0

# ── Data Acquisition ────────────────────────────────────────
roboflow>=1.1.0             # optional: Roboflow dataset download

# ── Development & Testing ───────────────────────────────────
pytest>=7.4.0
pytest-cov>=4.1.0

# ── Notebook Support ────────────────────────────────────────
jupyter>=1.0.0
ipywidgets>=8.0.0

# ── Annotation Tool (kept for labelling workflows) ──────────
labelImg>=1.8.6
